Fine tune OCALL opt-out granularity.
Previously, even if an enclave does not use a particular ocall,
the ocall will still exist in the host binary since the ocall implementation
was placed in the same file as some other symbol that the application (host) needs.

To solve this, OCALLS for some core system edls are placed into separate
c files whereever possible. This ensures that if an enclave does not need an ocall,
the implementation of the ocall will not be picked up.

Note: Weak Symbols (OE_WEAK_ALIAS) is not a viable solution since the linker will
still link in the weak symbol if it is placed in a file that contains another symbol
needed by the application.

fixes #3254
fixes #3255

Note: Not all system ocall implementations have been moved to separate C files.
