+++ host/sgx/ocalls/ocalls.h
@@ -4,7 +4,7 @@
 #ifndef _OE_HOST_SGX_OCALLS_H
 #define _OE_HOST_SGX_OCALLS_H
 
-#include "enclave.h"
+#include "../enclave.h"
 
 void HandleThreadWait(oe_enclave_t* enclave, uint64_t arg);
 void HandleThreadWake(oe_enclave_t* enclave, uint64_t arg);
